requirements.txt
amzn-sp-api==1.7.0
python-dotenv==1.0.0
